Reacomodar archivos compartidos en su propia carpeta de utils
